arc: some fixes

+++ files/board/arpl/overlayfs/opt/arpl/arc.sh
@@ -169,20 +169,37 @@ function arcMenu() {
   fi
   # Read model config for buildconfig
   NMODEL=${resp}
-  OMODEL=`printf "${NMODEL}" | jq -sRr @uri`
-  OURL="https://raw.githubusercontent.com/AuxXxilium/arc/main/files/board/arpl/overlayfs/opt/arpl/model-configs/${OMODEL}.yml"
-  if [ -f "${TMP_PATH}/${NMODEL}.yml" ]; then
-    rm -f "${TMP_PATH}/${NMODEL}.yml"
-  fi
-  OSTATUS="`curl --insecure -w "%{http_code}" -L "${OURL}" -o ${TMP_PATH}/${NMODEL}.yml`"
-  if [ $? -ne 0 -o ${OSTATUS} -ne 200 ]; then
-    dialog --backtitle "`backtitle`" --title "Online Config" --aspect 18 \
-      --infobox "No updated Modelconfig found" 0 0
-  else
-    dialog --backtitle "`backtitle`" --title "Online Config" --aspect 18 \
-      --infobox "Update Modelconfig to latest" 0 0
-    cp -f "${TMP_PATH}/${NMODEL}.yml" "${MODEL_CONFIG_PATH}/${NMODEL}.yml"
-  fi
+  while true; do
+    dialog --clear --backtitle "`backtitle`" \
+      --menu "Online Config" 0 0 0 \
+      1 "Update to latest Modelconfig" \
+      2 "Use local Modelconfig from Build" \
+    2>${TMP_PATH}/resp
+    [ $? -ne 0 ] && return
+    resp=$(<${TMP_PATH}/resp)
+    [ -z "${resp}" ] && return
+    if [ "${resp}" = "1" ]; then
+      OMODEL=`printf "${NMODEL}" | jq -sRr @uri`
+      OURL="https://raw.githubusercontent.com/AuxXxilium/arc/dev/files/board/arpl/overlayfs/opt/arpl/model-configs/${OMODEL}.yml"
+      if [ -f "${TMP_PATH}/${NMODEL}.yml" ]; then
+        rm -f "${TMP_PATH}/${NMODEL}.yml"
+      fi
+      OSTATUS="`curl --insecure -w "%{http_code}" -L "${OURL}" -o ${TMP_PATH}/${NMODEL}.yml`"
+      if [ $? -ne 0 -o ${OSTATUS} -ne 200 ]; then
+        dialog --backtitle "`backtitle`" --title "Online Config" --aspect 18 \
+          --infobox "No updated Modelconfig found!" 0 0
+      else
+        cp -f "${TMP_PATH}/${NMODEL}.yml" "${MODEL_CONFIG_PATH}/${NMODEL}.yml"
+        dialog --backtitle "`backtitle`" --title "Online Config" --aspect 18 \
+          --infobox "Updated Modelconfig to latest" 0 0
+      fi
+      break
+    elif [ "${resp}" = "2" ]; then
+      dialog --backtitle "`backtitle`" --title "Online Config" --aspect 18 \
+        --infobox "Use local Modelconfig from Build" 0 0
+      break
+    fi
+  done
   sleep 2
   if [ "${MODEL}" != "${NMODEL}" ]; then
     MODEL=${NMODEL}
